feat(formulas): Dedicated formulas language editor module
