+++ src/lib/ui/app/san-formulas/language/tokens-provider.ts
@@ -1,0 +1,67 @@
+import { languages } from 'monaco-editor/esm/vs/editor/editor.api.js'
+
+import { LANGUAGE_ID } from './language.js'
+import { DEFINITIONS } from './definitions.js'
+
+// NOTE: https://microsoft.github.io/monaco-editor/monarch.html
+languages.setMonarchTokensProvider(LANGUAGE_ID, {
+  //defaultToken: '',
+
+  keywords: DEFINITIONS.map((item) => item.label),
+
+  operators: ['+', '-', '*', '/', '=', ','],
+
+  // we include these common regular expressions
+  symbols: /[=><!~?:&|+\-*\/\^%]+/,
+
+  // C# style strings
+  escapes: /\\(?:[abfnrtv\\"']|x[0-9A-Fa-f]{1,4}|u[0-9A-Fa-f]{4}|U[0-9A-Fa-f]{8})/,
+
+  // The main tokenizer for our languages
+  tokenizer: {
+    root: [
+      // identifiers and keywords
+      [
+        /[a-z_$][\w$]*/,
+        {
+          cases: {
+            //'@typeKeywords': 'keyword',
+            '@keywords': 'keyword',
+            '@default': 'identifier',
+          },
+        },
+      ],
+
+      // whitespace
+      { include: '@whitespace' },
+
+      // delimiters and operators
+      [/[{}()\[\]]/, '@brackets'],
+      [/[<>](?!@symbols)/, '@brackets'],
+      [/@symbols/, { cases: { '@operators': 'operator', '@default': '' } }],
+
+      // numbers
+      [/\d*\.\d+([eE][\-+]?\d+)?/, 'number.float'],
+      [/\d+/, 'number'],
+
+      // delimiter: after number because of .\d floats
+      [/[;,.]/, 'delimiter'],
+
+      // strings
+      [/"([^"\\]|\\.)*$/, 'string.invalid'], // non-teminated string
+      [/"/, { token: 'string.quote', bracket: '@open', next: '@string' }],
+    ],
+
+    whitespace: [
+      [/[ \t\r\n]+/, 'white'],
+      [/#.*$/, 'comment'], // single line comment
+    ],
+
+    string: [
+      [/[^\\"]+/, 'string'],
+      [/@escapes/, 'string.escape'],
+      [/\\./, 'string.escape.invalid'],
+      [/"/, { token: 'string.quote', bracket: '@close', next: '@pop' }],
+    ],
+  },
+})
